Port Nathan Freeman's fix to contrib repository:
Fix for multisite redirect. Issue #74902

The multisite_redirect module threw an error when trying to save a
redirect.
There was also a place where a hook was commented out, probably because
it was broken for the same reason. I uncommented it because it is
needed for the redirects to behave as expected when edited.